% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalMove.R
\name{directionalMove}
\alias{directionalMove}
\title{This function models movement of lobsters toward the trap.The distance of lobsters to trap determines the magnitude of those moves.
As lobster gets closer to the trap, the magnitude of its directional move becomes larger and the random move becomes smaller.}
\usage{
directionalMove(
  Lobster,
  dStep,
  minDistoTrap,
  Trap,
  radiusOfInfluence,
  currentZoI
)
}
\arguments{
\item{Lobster}{location of lobster in the grid in x and y coordinates.}

\item{dStep}{Distance that each lobster moves during one time step.}

\item{minDistoTrap}{Distance from the trap.}

\item{Trap}{location of trap in the arena.}

\item{radiusOfInfluence}{Radius of influence for the baited trap.}

\item{currentZoI}{Radius of influence in each time step given the bait shrinkage.}
}
\value{
Returns the new coordinates of each lobster in the arena after each directional move.
}
\description{
This function models movement of lobsters toward the trap.The distance of lobsters to trap determines the magnitude of those moves.
As lobster gets closer to the trap, the magnitude of its directional move becomes larger and the random move becomes smaller.
}
