% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LobsterSizeFreqs}
\alias{LobsterSizeFreqs}
\title{Lobster size frequency data}
\format{
A data frame with 31 rows and 2 variables
}
\usage{
data(LobsterSizeFreqs)
}
\description{
The dataset contains frequency of each size bin (from Carapace length of 50 mm to 200 mm)
}
\details{
\itemize{
\item bins (Size groups/bins)
\item freq (Frequency)
}
}
\keyword{datasets}
