% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpoisD.R
\name{rpoisD}
\alias{rpoisD}
\title{This function generates  a Poisson or a negative binomial distribution for lobsters in the arena}
\usage{
rpoisD(n, lambda, D = 1)
}
\arguments{
\item{n}{is the number of lobsters to be generated}

\item{lambda}{is the mean density of lobsters}

\item{D}{is the dispersion index to be used. Default value is 1}
}
\value{
A vector of integers that is used as initial distribution of lobsters
}
\description{
This function generates  a Poisson or a negative binomial distribution for lobsters in the arena
}
