% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateLobsterMovement.R
\name{SimulateLobsterMovement}
\alias{SimulateLobsterMovement}
\title{Function to run the simulation based on defined parameters}
\usage{
SimulateLobsterMovement(p)
}
\arguments{
\item{p}{is a list of all input variables}
}
\value{
Returns a list
}
\description{
Function to run the simulation based on defined parameters
}
\seealso{
Examples of the input parameters can be found here: https://github.com/vpourfaraj/lobsterCatch/tree/main/R
}
