% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapInPath.R
\name{trapInPath}
\alias{trapInPath}
\title{This function determines if lobster gets into a trap and is caught.}
\usage{
trapInPath(loc1, loc2, Trap, howClose)
}
\arguments{
\item{loc1}{is the location of lobster at the start of each time step}

\item{loc2}{is the location of lobster at the end of each time step}

\item{Trap}{is the location of trap}

\item{howClose}{The area within which a lobster considered trapped}
}
\value{
Returns a vector that contain lobster path and whether its trapped
}
\description{
This function determines if lobster gets into a trap and is caught.
}
