% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directionalMove.R
\name{directionalMove}
\alias{directionalMove}
\title{This function controls the movements of lobsters toward the trap. There are both directional and random move components considered in
this function. The proximity of a lobster to a trap controls the magnitude of each components, i.e closer a lobster gets to the trap
the smaller the random component of movement. The directional move depends on radius of influence of a trap and current zone of
influence.}
\usage{
directionalMove(
  Lobster,
  dStep,
  minDistoTrap,
  Trap,
  radiusOfInfluence,
  currentZoI
)
}
\arguments{
\item{Lobster}{location of lobster in the grid in x and y coordinates.}

\item{dStep}{defines the distance that each lobster moves in each time step.}

\item{minDistoTrap}{defines the minimum distance to the trap.}

\item{Trap}{location of trap in the grid in x and y coordinates.}

\item{radiusOfInfluence}{trap's radius of influence.}

\item{currentZoI}{initial zone of influence}
}
\value{
Returns the new coordinates of each lobster in the grid after a directional move is applied.
}
\description{
This function controls the movements of lobsters toward the trap. There are both directional and random move components considered in
this function. The proximity of a lobster to a trap controls the magnitude of each components, i.e closer a lobster gets to the trap
the smaller the random component of movement. The directional move depends on radius of influence of a trap and current zone of
influence.
}
