% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceToClosestTrap.R
\name{distanceToClosestTrap}
\alias{distanceToClosestTrap}
\title{The function finds the closest trap to a lobster and calculates the distance.}
\usage{
distanceToClosestTrap(Lobster, Trap)
}
\arguments{
\item{Lobster}{location of lobster in the arena}

\item{Trap}{location of trap in the arena}
}
\value{
Returns distance to closest trap and saves the trap number in case of multiple traps.
}
\description{
The function finds the closest trap to a lobster and calculates the distance.
}
