% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSimOutput.R
\name{GetSimOutput}
\alias{GetSimOutput}
\title{This function extracts the results of simulation.}
\usage{
GetSimOutput(x, mls = 82.5)
}
\arguments{
\item{x}{is an object generated by \code{SimulateLobsterMovement} function.}

\item{mls}{is the minimum legal size(mls) in mm. The default is 82.5 mm.}
}
\value{
Returns the followings for each replicate: the number of lobsters caught,
legal catch weight (bigger than mls), total catch weight and length of time to reach maximum catch.
}
\description{
This function extracts the results of simulation.
}
