% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LobsterSizeFreqs}
\alias{LobsterSizeFreqs}
\title{Dataset of bins and freq}
\format{
A data frame with 31 rows and 2 variables
}
\usage{
data(LobsterSizeFreqs)
}
\description{
to-do-list: describe dataset here
}
\details{
\itemize{
\item bins (DESCRIPTION GOES HERE)
\item freq (DESCRIPTION GOES HERE)
}
}
\keyword{datasets}
